Finished all Spot routes, and Review routes, confirmed working and erroring correctly

+++ backend/db/migrations/20230609031221-create-review.js
@@ -34,7 +34,7 @@ module.exports = {
           type: Sequelize.STRING,
         },
         stars: {
-          type: Sequelize.DECIMAL,
+          type: Sequelize.INTEGER,
         },
         createdAt: {
           allowNull: false,
